Customer recommendation to market towards
Historical data set (training.csv) of customers. As well as a list of potential customers to whom to market (testingCandidate.csv). From this list of potential customers, I determine whether it is profitable to market to them. The cost of marketing to a particular customer is $30 and the cost is paid regardless of whether the customer responds to the marketing or not.